+++ modules/workflow-engine/src/main/java/io/mateu/workflow/infra/config/WorkflowEngineConfiguration.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableConfigurationProperties({GitImportProperties.class, MessageApiProperties.class})
+@EnableConfigurationProperties({GitImportProperties.class, DirectoryImportProperties.class, MessageApiProperties.class})
 public class WorkflowEngineConfiguration {
 
     /**